WIP - error messages and code cleanup
- plus example permission config

#5143

+++ waltz-service/src/main/java/com/khartec/waltz/service/custom_environment/CustomEnvironmentUsageService.java
@@ -1,5 +1,6 @@
 package com.khartec.waltz.service.custom_environment;
 
+import com.khartec.waltz.common.exception.InsufficientPrivelegeException;
 import com.khartec.waltz.data.custom_environment.CustomEnvironmentUsageDao;
 import com.khartec.waltz.model.EntityKind;
 import com.khartec.waltz.model.EntityReference;
@@ -11,6 +12,7 @@ import com.khartec.waltz.model.custom_environment.CustomEnvironment;
 import com.khartec.waltz.model.custom_environment.CustomEnvironmentUsage;
 import com.khartec.waltz.model.custom_environment.CustomEnvironmentUsageInfo;
 import com.khartec.waltz.service.changelog.ChangeLogService;
+import com.khartec.waltz.service.permission.PermissionGroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,14 +27,17 @@ public class CustomEnvironmentUsageService {
     private final CustomEnvironmentUsageDao customEnvironmentUsageDao;
     private final CustomEnvironmentService customEnvironmentService;
     private final ChangeLogService changeLogService;
+    private final PermissionGroupService permissionGroupService;
 
     @Autowired
     public CustomEnvironmentUsageService(CustomEnvironmentUsageDao customEnvironmentUsageDao,
                                          CustomEnvironmentService customEnvironmentService,
-                                         ChangeLogService changeLogService){
+                                         ChangeLogService changeLogService,
+                                         PermissionGroupService permissionGroupService){
         this.customEnvironmentUsageDao = customEnvironmentUsageDao;
         this.customEnvironmentService = customEnvironmentService;
         this.changeLogService = changeLogService;
+        this.permissionGroupService = permissionGroupService;
     }
 
 
@@ -41,29 +46,40 @@ public class CustomEnvironmentUsageService {
     }
 
 
-    public Long addAsset(CustomEnvironmentUsage usage, String username){
-        Long usageId = customEnvironmentUsageDao.addAsset(usage, username);
-
-        CustomEnvironmentUsageInfo usageInfo = customEnvironmentUsageDao.getUsageInfoById(usageId);
+    public Long addAsset(CustomEnvironmentUsage usage, String username) throws InsufficientPrivelegeException {
         CustomEnvironment customEnvironment = customEnvironmentService.getById(usage.customEnvironmentId());
-        String message = format("Added asset: %s to custom environment: %s/%s",
+
+        ensureUserHasPermission(customEnvironment, username, Operation.ADD);
+
+        Long usageId = customEnvironmentUsageDao.addAsset(usage, username);
+        CustomEnvironmentUsageInfo usageInfo = customEnvironmentUsageDao.getUsageInfoById(usageId);
+        String message = format(
+                "Added asset: %s to custom environment: %s/%s",
                 pretty(usageInfo.asset().entityReference()),
                 customEnvironment.groupName(),
                 customEnvironment.name());
 
-        ChangeLog changeLog = mkChangeLog(customEnvironment.owningEntity(), username, message, Operation.ADD);
+        ChangeLog changeLog = mkChangeLog(
+                customEnvironment.owningEntity(),
+                username,
+                message,
+                Operation.ADD);
+
         changeLogService.write(changeLog);
 
         return usageId;
     }
 
 
-    public Boolean remove(Long usageId, String username){
+    public Boolean remove(Long usageId, String username) throws InsufficientPrivelegeException {
         CustomEnvironmentUsageInfo usageInfo = customEnvironmentUsageDao.getUsageInfoById(usageId);
+        CustomEnvironment customEnvironment = customEnvironmentService.getById(usageInfo.usage().customEnvironmentId());
+
+        ensureUserHasPermission(customEnvironment, username, Operation.REMOVE);
+
         boolean remove = customEnvironmentUsageDao.remove(usageId);
 
         if(remove) {
-            CustomEnvironment customEnvironment = customEnvironmentService.getById(usageInfo.usage().customEnvironmentId());
             String message = format("Removed asset: %s from custom environment: %s/%s",
                     pretty(usageInfo.asset().entityReference()),
                     customEnvironment.groupName(),
@@ -94,6 +110,23 @@ public class CustomEnvironmentUsageService {
                 .childKind(EntityKind.CUSTOM_ENVIRONMENT_USAGE)
                 .operation(operation)
                 .build();
+    }
+
+
+    private void ensureUserHasPermission(CustomEnvironment env,
+                                         String username,
+                                         Operation op) throws InsufficientPrivelegeException {
+        boolean hasPerm = permissionGroupService.hasPermission(
+                env.owningEntity(),
+                EntityKind.CUSTOM_ENVIRONMENT,
+                username);
+
+        if (!hasPerm) {
+            String msg = format(
+                    "Cannot %s environment usage, insufficient permissions",
+                    op.name().toLowerCase());
+            throw new InsufficientPrivelegeException(msg);
+        }
     }
 
 }
